Renamed "Task" -> "Goal"
The Core Data entity was named Task. This conflicts with the built in Swift type Task, which means we can't write concurrent code.

For this reason I renamed Task. Since Goal is used throughout the UI, I renamed Task to Goal as well as any other Views and functions, in order to keep consistency.